Move kas_wgpu::theme contents to kas_theme
