+++ kas-theme/src/flat_theme.rs
@@ -7,29 +7,27 @@
 //!
 //! Widget size and appearance can be modified through themes.
 
+use rusttype::Font;
 use std::f32;
-use wgpu_glyph::Font;
 
-use kas::draw::{Colour, Draw, DrawShaded, DrawText};
+use crate::{Dimensions, DimensionsParams, DimensionsWindow, Theme, ThemeColours};
+use kas::draw::{Colour, Draw, DrawText};
 use kas::event::HighlightState;
 use kas::geom::{Coord, Rect};
 use kas::theme::{self, TextClass, TextProperties, ThemeAction, ThemeApi};
 use kas::Direction;
-use kas_theme::{Dimensions, DimensionsParams, DimensionsWindow, Theme, ThemeColours};
 
-use crate::draw::DrawPipe;
-
-/// A simple, inflexible theme providing a sample implementation.
+/// A simple flat theme.
 #[derive(Clone, Debug)]
-pub struct ShadedTheme {
+pub struct FlatTheme {
     font_size: f32,
     cols: ThemeColours,
 }
 
-impl ShadedTheme {
+impl FlatTheme {
     /// Construct
     pub fn new() -> Self {
-        ShadedTheme {
+        FlatTheme {
             font_size: 18.0,
             cols: ThemeColours::new(),
         }
@@ -38,29 +36,29 @@ impl ShadedTheme {
 
 const DIMS: DimensionsParams = DimensionsParams {
     margin: 2.0,
-    frame_size: 5.0,
-    button_frame: 5.0,
+    frame_size: 4.0,
+    button_frame: 6.0,
     scrollbar_size: 8.0,
 };
 
-pub struct DrawHandle<'a> {
-    draw: &'a mut DrawPipe,
+pub struct DrawHandle<'a, D: Draw> {
+    draw: &'a mut D,
     window: &'a mut DimensionsWindow,
     cols: &'a ThemeColours,
     rect: Rect,
     offset: Coord,
-    pass: usize,
+    pass: <D as Draw>::Region,
 }
 
-impl Theme<DrawPipe> for ShadedTheme {
+impl<D: Draw + DrawText + 'static> Theme<D> for FlatTheme {
     type Window = DimensionsWindow;
 
     #[cfg(not(feature = "gat"))]
-    type DrawHandle = DrawHandle<'static>;
+    type DrawHandle = DrawHandle<'static, D>;
     #[cfg(feature = "gat")]
-    type DrawHandle<'a> = DrawHandle<'a>;
+    type DrawHandle<'a> = DrawHandle<'a, D>;
 
-    fn new_window(&self, _draw: &mut DrawPipe, dpi_factor: f32) -> Self::Window {
+    fn new_window(&self, _draw: &mut D, dpi_factor: f32) -> Self::Window {
         DimensionsWindow::new(DIMS, self.font_size, dpi_factor)
     }
 
@@ -71,25 +69,25 @@ impl Theme<DrawPipe> for ShadedTheme {
     #[cfg(not(feature = "gat"))]
     unsafe fn draw_handle<'a>(
         &'a self,
-        draw: &'a mut DrawPipe,
+        draw: &'a mut D,
         window: &'a mut Self::Window,
         rect: Rect,
     ) -> Self::DrawHandle {
         // We extend lifetimes (unsafe) due to the lack of associated type generics.
         use std::mem::transmute;
         DrawHandle {
-            draw: transmute::<&'a mut DrawPipe, &'static mut DrawPipe>(draw),
+            draw: transmute::<&'a mut D, &'static mut D>(draw),
             window: transmute::<&'a mut Self::Window, &'static mut Self::Window>(window),
             cols: transmute::<&'a ThemeColours, &'static ThemeColours>(&self.cols),
             rect,
             offset: Coord::ZERO,
-            pass: 0,
+            pass: <D as Draw>::Region::default(),
         }
     }
     #[cfg(feature = "gat")]
     fn draw_handle<'a>(
         &'a self,
-        draw: &'a mut DrawPipe,
+        draw: &'a mut D,
         window: &'a mut Self::Window,
         rect: Rect,
     ) -> Self::DrawHandle<'a> {
@@ -99,12 +97,12 @@ impl Theme<DrawPipe> for ShadedTheme {
             cols: &self.cols,
             rect,
             offset: Coord::ZERO,
-            pass: 0,
+            pass: <D as Draw>::Region::default(),
         }
     }
 
     fn get_fonts<'a>(&self) -> Vec<Font<'a>> {
-        vec![kas_theme::get_font()]
+        vec![crate::get_font()]
     }
 
     fn light_direction(&self) -> (f32, f32) {
@@ -116,7 +114,7 @@ impl Theme<DrawPipe> for ShadedTheme {
     }
 }
 
-impl ThemeApi for ShadedTheme {
+impl ThemeApi for FlatTheme {
     fn set_font_size(&mut self, size: f32) -> ThemeAction {
         self.font_size = size;
         ThemeAction::ThemeResize
@@ -132,26 +130,29 @@ impl ThemeApi for ShadedTheme {
     }
 }
 
-impl<'a> DrawHandle<'a> {
+impl<'a, D: Draw> DrawHandle<'a, D> {
     /// Draw an edit region with optional navigation highlight.
     /// Return the inner rect.
-    fn draw_edit_region(&mut self, mut outer: Rect, nav_col: Option<Colour>) -> Rect {
-        let mut inner = outer.shrink(self.window.dims.frame);
+    fn draw_edit_region(&mut self, outer: Rect, nav_col: Option<Colour>) -> Rect {
+        let inner1 = outer.shrink(self.window.dims.frame / 2);
+        let inner2 = outer.shrink(self.window.dims.frame);
+
+        self.draw.rect(self.pass, inner1, self.cols.text_area);
+
+        // We draw over the inner rect, taking advantage of the fact that
+        // rounded frames get drawn after flat rects.
         self.draw
-            .shaded_square_frame(self.pass, outer, inner, (-0.8, 0.0), self.cols.background);
+            .rounded_frame(self.pass, outer, inner2, 0.333, self.cols.frame);
 
         if let Some(col) = nav_col {
-            outer = inner;
-            inner = outer.shrink(self.window.dims.margin);
-            self.draw.frame(self.pass, outer, inner, col);
+            self.draw.rounded_frame(self.pass, inner1, inner2, 0.0, col);
         }
 
-        self.draw.rect(self.pass, inner, self.cols.text_area);
-        inner
+        inner2
     }
 }
 
-impl<'a> theme::DrawHandle for DrawHandle<'a> {
+impl<'a, D: Draw + DrawText> theme::DrawHandle for DrawHandle<'a, D> {
     fn clip_region(
         &mut self,
         rect: Rect,
@@ -180,7 +181,7 @@ impl<'a> theme::DrawHandle for DrawHandle<'a> {
         let outer = rect + self.offset;
         let inner = outer.shrink(self.window.dims.frame);
         self.draw
-            .shaded_round_frame(self.pass, outer, inner, (0.6, -0.6), self.cols.background);
+            .rounded_frame(self.pass, outer, inner, 0.5, self.cols.frame);
     }
 
     fn text(&mut self, rect: Rect, text: &str, props: TextProperties) {
@@ -195,11 +196,10 @@ impl<'a> theme::DrawHandle for DrawHandle<'a> {
 
     fn button(&mut self, rect: Rect, highlights: HighlightState) {
         let outer = rect + self.offset;
-        let inner = outer.shrink(self.window.dims.button_frame);
         let col = self.cols.button_state(highlights);
 
-        self.draw
-            .shaded_round_frame(self.pass, outer, inner, (0.0, 0.6), col);
+        let inner = outer.shrink(self.window.dims.button_frame);
+        self.draw.rounded_frame(self.pass, outer, inner, 0.0, col);
         self.draw.rect(self.pass, inner, col);
 
         if let Some(col) = self.cols.nav_region(highlights) {
@@ -224,7 +224,14 @@ impl<'a> theme::DrawHandle for DrawHandle<'a> {
         let inner = self.draw_edit_region(rect + self.offset, nav_col);
 
         if let Some(col) = self.cols.check_mark_state(highlights, checked) {
-            self.draw.shaded_square(self.pass, inner, (0.0, 0.4), col);
+            let radius = (inner.size.0 + inner.size.1) / 16;
+            let inner = inner.shrink(self.window.dims.margin + radius);
+            let p1 = inner.pos;
+            let p2 = inner.pos + inner.size;
+            let radius = radius as f32;
+            self.draw.rounded_line(self.pass, p1, p2, radius, col);
+            self.draw
+                .rounded_line(self.pass, Coord(p1.0, p2.1), Coord(p2.0, p1.1), radius, col);
         }
     }
 
@@ -241,7 +248,8 @@ impl<'a> theme::DrawHandle for DrawHandle<'a> {
         let inner = self.draw_edit_region(rect + self.offset, nav_col);
 
         if let Some(col) = self.cols.check_mark_state(highlights, checked) {
-            self.draw.shaded_circle(self.pass, inner, (0.0, 1.0), col);
+            let inner = inner.shrink(self.window.dims.margin);
+            self.draw.circle(self.pass, inner, 0.3, col);
         }
     }
 
@@ -258,8 +266,7 @@ impl<'a> theme::DrawHandle for DrawHandle<'a> {
         let half_width = outer.size.0.min(outer.size.1) / 2;
         let inner = outer.shrink(half_width);
         let col = self.cols.scrollbar_state(highlights);
-        self.draw
-            .shaded_round_frame(self.pass, outer, inner, (0.0, 0.6), col);
+        self.draw.rounded_frame(self.pass, outer, inner, 0.0, col);
         self.draw.rect(self.pass, inner, col);
     }
 }
