+++ kas-theme/src/shaded_theme.rs
@@ -7,29 +7,27 @@
 //!
 //! Widget size and appearance can be modified through themes.
 
+use rusttype::Font;
 use std::f32;
-use wgpu_glyph::Font;
 
-use kas::draw::{Colour, Draw, DrawText};
+use crate::{Dimensions, DimensionsParams, DimensionsWindow, Theme, ThemeColours};
+use kas::draw::{Colour, Draw, DrawShaded, DrawText};
 use kas::event::HighlightState;
 use kas::geom::{Coord, Rect};
 use kas::theme::{self, TextClass, TextProperties, ThemeAction, ThemeApi};
 use kas::Direction;
-use kas_theme::{Dimensions, DimensionsParams, DimensionsWindow, Theme, ThemeColours};
 
-use crate::draw::DrawPipe;
-
-/// A simple flat theme.
+/// A simple, inflexible theme providing a sample implementation.
 #[derive(Clone, Debug)]
-pub struct FlatTheme {
+pub struct ShadedTheme {
     font_size: f32,
     cols: ThemeColours,
 }
 
-impl FlatTheme {
+impl ShadedTheme {
     /// Construct
     pub fn new() -> Self {
-        FlatTheme {
+        ShadedTheme {
             font_size: 18.0,
             cols: ThemeColours::new(),
         }
@@ -38,29 +36,32 @@ impl FlatTheme {
 
 const DIMS: DimensionsParams = DimensionsParams {
     margin: 2.0,
-    frame_size: 4.0,
-    button_frame: 6.0,
+    frame_size: 5.0,
+    button_frame: 5.0,
     scrollbar_size: 8.0,
 };
 
-pub struct DrawHandle<'a> {
-    draw: &'a mut DrawPipe,
+pub struct DrawHandle<'a, D: Draw> {
+    draw: &'a mut D,
     window: &'a mut DimensionsWindow,
     cols: &'a ThemeColours,
     rect: Rect,
     offset: Coord,
-    pass: usize,
+    pass: <D as Draw>::Region,
 }
 
-impl Theme<DrawPipe> for FlatTheme {
+impl<D> Theme<D> for ShadedTheme
+where
+    D: Draw + DrawShaded + DrawText + 'static,
+{
     type Window = DimensionsWindow;
 
     #[cfg(not(feature = "gat"))]
-    type DrawHandle = DrawHandle<'static>;
+    type DrawHandle = DrawHandle<'static, D>;
     #[cfg(feature = "gat")]
-    type DrawHandle<'a> = DrawHandle<'a>;
+    type DrawHandle<'a> = DrawHandle<'a, D>;
 
-    fn new_window(&self, _draw: &mut DrawPipe, dpi_factor: f32) -> Self::Window {
+    fn new_window(&self, _draw: &mut D, dpi_factor: f32) -> Self::Window {
         DimensionsWindow::new(DIMS, self.font_size, dpi_factor)
     }
 
@@ -71,25 +72,25 @@ impl Theme<DrawPipe> for FlatTheme {
     #[cfg(not(feature = "gat"))]
     unsafe fn draw_handle<'a>(
         &'a self,
-        draw: &'a mut DrawPipe,
+        draw: &'a mut D,
         window: &'a mut Self::Window,
         rect: Rect,
     ) -> Self::DrawHandle {
         // We extend lifetimes (unsafe) due to the lack of associated type generics.
         use std::mem::transmute;
         DrawHandle {
-            draw: transmute::<&'a mut DrawPipe, &'static mut DrawPipe>(draw),
+            draw: transmute::<&'a mut D, &'static mut D>(draw),
             window: transmute::<&'a mut Self::Window, &'static mut Self::Window>(window),
             cols: transmute::<&'a ThemeColours, &'static ThemeColours>(&self.cols),
             rect,
             offset: Coord::ZERO,
-            pass: 0,
+            pass: <D as Draw>::Region::default(),
         }
     }
     #[cfg(feature = "gat")]
     fn draw_handle<'a>(
         &'a self,
-        draw: &'a mut DrawPipe,
+        draw: &'a mut D,
         window: &'a mut Self::Window,
         rect: Rect,
     ) -> Self::DrawHandle<'a> {
@@ -99,12 +100,12 @@ impl Theme<DrawPipe> for FlatTheme {
             cols: &self.cols,
             rect,
             offset: Coord::ZERO,
-            pass: 0,
+            pass: <D as Draw>::Region::default(),
         }
     }
 
     fn get_fonts<'a>(&self) -> Vec<Font<'a>> {
-        vec![kas_theme::get_font()]
+        vec![crate::get_font()]
     }
 
     fn light_direction(&self) -> (f32, f32) {
@@ -116,7 +117,7 @@ impl Theme<DrawPipe> for FlatTheme {
     }
 }
 
-impl ThemeApi for FlatTheme {
+impl ThemeApi for ShadedTheme {
     fn set_font_size(&mut self, size: f32) -> ThemeAction {
         self.font_size = size;
         ThemeAction::ThemeResize
@@ -132,29 +133,26 @@ impl ThemeApi for FlatTheme {
     }
 }
 
-impl<'a> DrawHandle<'a> {
+impl<'a, D: Draw + DrawShaded> DrawHandle<'a, D> {
     /// Draw an edit region with optional navigation highlight.
     /// Return the inner rect.
-    fn draw_edit_region(&mut self, outer: Rect, nav_col: Option<Colour>) -> Rect {
-        let inner1 = outer.shrink(self.window.dims.frame / 2);
-        let inner2 = outer.shrink(self.window.dims.frame);
-
-        self.draw.rect(self.pass, inner1, self.cols.text_area);
-
-        // We draw over the inner rect, taking advantage of the fact that
-        // rounded frames get drawn after flat rects.
+    fn draw_edit_region(&mut self, mut outer: Rect, nav_col: Option<Colour>) -> Rect {
+        let mut inner = outer.shrink(self.window.dims.frame);
         self.draw
-            .rounded_frame(self.pass, outer, inner2, 0.333, self.cols.frame);
+            .shaded_square_frame(self.pass, outer, inner, (-0.8, 0.0), self.cols.background);
 
         if let Some(col) = nav_col {
-            self.draw.rounded_frame(self.pass, inner1, inner2, 0.0, col);
+            outer = inner;
+            inner = outer.shrink(self.window.dims.margin);
+            self.draw.frame(self.pass, outer, inner, col);
         }
 
-        inner2
+        self.draw.rect(self.pass, inner, self.cols.text_area);
+        inner
     }
 }
 
-impl<'a> theme::DrawHandle for DrawHandle<'a> {
+impl<'a, D: Draw + DrawShaded + DrawText> theme::DrawHandle for DrawHandle<'a, D> {
     fn clip_region(
         &mut self,
         rect: Rect,
@@ -183,7 +181,7 @@ impl<'a> theme::DrawHandle for DrawHandle<'a> {
         let outer = rect + self.offset;
         let inner = outer.shrink(self.window.dims.frame);
         self.draw
-            .rounded_frame(self.pass, outer, inner, 0.5, self.cols.frame);
+            .shaded_round_frame(self.pass, outer, inner, (0.6, -0.6), self.cols.background);
     }
 
     fn text(&mut self, rect: Rect, text: &str, props: TextProperties) {
@@ -198,10 +196,11 @@ impl<'a> theme::DrawHandle for DrawHandle<'a> {
 
     fn button(&mut self, rect: Rect, highlights: HighlightState) {
         let outer = rect + self.offset;
+        let inner = outer.shrink(self.window.dims.button_frame);
         let col = self.cols.button_state(highlights);
 
-        let inner = outer.shrink(self.window.dims.button_frame);
-        self.draw.rounded_frame(self.pass, outer, inner, 0.0, col);
+        self.draw
+            .shaded_round_frame(self.pass, outer, inner, (0.0, 0.6), col);
         self.draw.rect(self.pass, inner, col);
 
         if let Some(col) = self.cols.nav_region(highlights) {
@@ -226,14 +225,7 @@ impl<'a> theme::DrawHandle for DrawHandle<'a> {
         let inner = self.draw_edit_region(rect + self.offset, nav_col);
 
         if let Some(col) = self.cols.check_mark_state(highlights, checked) {
-            let radius = (inner.size.0 + inner.size.1) / 16;
-            let inner = inner.shrink(self.window.dims.margin + radius);
-            let p1 = inner.pos;
-            let p2 = inner.pos + inner.size;
-            let radius = radius as f32;
-            self.draw.rounded_line(self.pass, p1, p2, radius, col);
-            self.draw
-                .rounded_line(self.pass, Coord(p1.0, p2.1), Coord(p2.0, p1.1), radius, col);
+            self.draw.shaded_square(self.pass, inner, (0.0, 0.4), col);
         }
     }
 
@@ -250,8 +242,7 @@ impl<'a> theme::DrawHandle for DrawHandle<'a> {
         let inner = self.draw_edit_region(rect + self.offset, nav_col);
 
         if let Some(col) = self.cols.check_mark_state(highlights, checked) {
-            let inner = inner.shrink(self.window.dims.margin);
-            self.draw.circle(self.pass, inner, 0.3, col);
+            self.draw.shaded_circle(self.pass, inner, (0.0, 1.0), col);
         }
     }
 
@@ -268,7 +259,8 @@ impl<'a> theme::DrawHandle for DrawHandle<'a> {
         let half_width = outer.size.0.min(outer.size.1) / 2;
         let inner = outer.shrink(half_width);
         let col = self.cols.scrollbar_state(highlights);
-        self.draw.rounded_frame(self.pass, outer, inner, 0.0, col);
+        self.draw
+            .shaded_round_frame(self.pass, outer, inner, (0.0, 0.6), col);
         self.draw.rect(self.pass, inner, col);
     }
 }
