--- kas-wgpu/src/theme/mod.rs
@@ -1,13 +1,0 @@
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License in the LICENSE-APACHE file or at:
-//     https://www.apache.org/licenses/LICENSE-2.0
-
-//! Themes
-
-mod flat_theme;
-mod shaded_theme;
-
-pub use flat_theme::FlatTheme;
-pub use kas::theme::*;
-pub use shaded_theme::ShadedTheme;
